Refactor Bison and Flex files in bison_test folder

+++ liasm/includes/FirstPassAnalyzer.hpp
@@ -11,22 +11,22 @@ class FirstPassAnalyzer
 {
 public:
     FirstPassAnalyzer(
-        std::unordered_map<std::string, int> &memory,
-        std::unordered_map<std::string, int> &labels,
-        std::unordered_map<int, std::string> &program);
+        std::unordered_map<std::string, u_int16_t> &memory,
+        std::unordered_map<std::string, u_int16_t> &labels,
+        std::unordered_map<u_int16_t, std::string> &program);
 
     int analyse(std::ifstream &file);
     void printLabels();
 
 private:
-    std::unordered_map<std::string, int> &memory;
-    std::unordered_map<std::string, int> &labels;
-    std::unordered_map<int, std::string> &program;
+    std::unordered_map<std::string, u_int16_t> &memory;
+    std::unordered_map<std::string, u_int16_t> &labels;
+    std::unordered_map<u_int16_t, std::string> &program;
 
     bool stopFlag = false;
 
     void processInstruction(std::string &instr, std::istringstream &iss,
-                            std::string &label, int &address);
+                            std::string &label, u_int16_t &address);
 };
 
 #endif // FIRSTPASSANALYZER_HPP
